api: replace media list in Playlist resource by mediaForUser
The Playlist serialiser was previously doing a query each time it was
rendered which was inefficient. Replace this with annotating the
retrieved object with a single queryset which returns all of the media
items viewable by a user.

This also fixes a bug whereby a playlist resource's media list could
include media items invisible to the current user.