ui: add RSS view of playlist
Add a specialised view of a playlist as an RSS feed available under
/playlists/<pk>.rss. We have to make sure to use the "slug" pattern in
Django to match the playlist key as otherwise it greedily matches
"foo.rss" as the playlist id.

The specialist view is based on the playlist API view but with a custom
renderer which uses feedgen to generate the RSS feed. The feed respects
the "downloadable" property and makes sure that non-downloadable videos
do not appear in the list.