api: move get_queryset specialisation into ChannelMixin
The get_queryset specialisation in ChannelView was required for channels
to be serialised (unlike the one in MediaItemView which is only required
for filtering). Having it in the view and not the mixin meant that the
UI view did not annotate the correct information.

This was not found earlier because there was not a test for the channel
view in the UI so we add one.

Closes #339