oaipmh: add Track object
Add a new database object, Track, which represents an individual track
within a MatterhornRecord. Add logic to the MatterhornRecord creation
hook to create Track objects from the record as appropriate.

For the moment, we only populate the database with at most one track
from each record. The track must have the correct "type" attribute
(currently "presentation/delivery") to be copied to the database.

At most one track is copied for each record so if multiple tracks are
present, the first one in the document is used.

After this commit is applied, a full cleanup will be required to create
any tracks for existing records.

At this point, the database contains URLs to all videos which should be
created in the Track object along with destination playlists reachable
via Track -> MatterhornRecord -> Series.playlist. Media items created
for each track can be recorded in the track object.